TASK-6: Decomposed Exceptions.cs: moved each exception to a separate file, added to a new directory
