+++ TASK-6/Exceptions/PhoneNotFoundException.cs
@@ -1,0 +1,16 @@
+namespace TASK_6.Exceptions;
+
+public class PhoneNotFoundException : Exception
+{
+    public PhoneNotFoundException()
+    {
+    }
+
+    public PhoneNotFoundException(string message) : base(message)
+    {
+    }
+
+    public PhoneNotFoundException(string message, Exception inner) : base(message, inner)
+    {
+    }
+}
