+++ TASK-6/Exceptions/StoreNotFoundException.cs
@@ -1,0 +1,16 @@
+namespace TASK_6.Exceptions;
+
+public class StoreNotFoundException : Exception
+{
+    public StoreNotFoundException()
+    {
+    }
+
+    public StoreNotFoundException(string message) : base(message)
+    {
+    }
+
+    public StoreNotFoundException(string message, Exception inner) : base(message, inner)
+    {
+    }
+}
