+++ TASK-6/JsonShops.cs
@@ -1,4 +1,5 @@
 using NLog;
+using TASK_6.Exceptions;
 
 namespace TASK_6;
